Module React

The React library

type domRef

type 'value ref = {

mutable current : 'value;

}

module Ref : sig ... end

val createRef : unit -> 'a option ref

val useRef : 'a -> 'a ref

val forwardRef : (unit -> 'a) -> 'a

module Event : sig ... end

module JSX : sig ... end

All of those types are used by the server-reason-react.ppx internally to represent valid React code from the server. It currently different from reason-react-ppx due to a need for knowing the types since ReactDOM needs to render differently depending on the type.

type error = {

message : string;

stack : Yojson.Basic.t;

env : string;

digest : string;

}

module Model : sig ... end

type element =

| Lower_case_element of lower_case_element

| Upper_case_component of string * unit -> element

| Async_component of string * unit -> element Lwt.t

| Client_component of {

props : client_props;

client : element;

import_module : string;

import_name : string;

}

| List of element list

| Array of element array

| Text of string

| Static of {

prerendered : string;

original : element;

}

| Fragment of element

| Empty

| Provider of {

children : element;

push : unit -> unit -> unit;

async_key : Obj.t Lwt.key;

async_value : Obj.t;

}

| Consumer of element

| Suspense of {

key : string option;

children : element;

fallback : element;

}

and lower_case_element = {

key : string option;

tag : string;

attributes : JSX.prop list;

children : element list;

}

and client_props = (string * element Model.t) list

and model_value = element Model.t

exception Invalid_children of string

module Fragment : sig ... end

val createElement : string -> JSX.prop list -> element list -> element

val createElementWithKey : ?key:string -> string -> JSX.prop list -> element list -> element

val fragment : element -> element

val cloneElement : element -> JSX.prop list -> element

val string : string -> element

val null : element

val int : int -> element

val float : float -> element

val array : element array -> element

val list : element list -> element

type 'a provider = value:'a -> children:element -> unit -> element

module Context : sig ... end

val createContext : 'a -> 'a Context.t

module Suspense : sig ... end

module Cache : sig ... end

type any_promise =

| Any_promise : 'a Lwt.t -> any_promise

exception Suspend of any_promise

val memo : (('props * 'props) -> bool) -> 'a -> ('props * 'props) -> bool

val memoCustomCompareProps : (('props * 'props) -> bool) -> (('props * 'props) -> bool) -> 'a -> ('props * 'props) -> bool

val cache : ('a -> 'b) -> 'a -> 'b

val useContext : 'a Context.t -> 'a

val useState : (unit -> 'state) -> 'state * (('state -> 'state) -> unit)

val useMemo : (unit -> 'a) -> 'a

val useMemo0 : (unit -> 'a) -> 'a

val useMemo1 : (unit -> 'a) -> 'b -> 'a

val useMemo2 : (unit -> 'a) -> 'b -> 'a

val useMemo3 : (unit -> 'a) -> 'b -> 'a

val useMemo4 : (unit -> 'a) -> 'b -> 'a

val useMemo5 : (unit -> 'a) -> 'b -> 'a

val useMemo6 : (unit -> 'a) -> 'b -> 'a

val useCallback : 'a -> 'a

val useCallback0 : 'a -> 'a

val useCallback1 : 'a -> 'b -> 'a

val useCallback2 : 'a -> 'b -> 'a

val useCallback3 : 'a -> 'b -> 'a

val useCallback4 : 'a -> 'b -> 'a

val useCallback5 : 'a -> 'b -> 'a

val useCallback6 : 'a -> 'b -> 'a

val useId : unit -> string

type ('input, 'output) callback = 'input -> 'output

val useSyncExternalStore : subscribe:((unit -> unit) -> (unit, unit) callback) -> getSnapshot:(unit -> 'snapshot) -> 'snapshot

deprecated Use useSyncExternalStoreWithServer instead

val useSyncExternalStoreWithServer : subscribe:((unit -> unit) -> (unit, unit) callback) -> getSnapshot:(unit -> 'snapshot) -> getServerSnapshot:(unit -> 'snapshot) -> 'snapshot

val useReducer : ('state -> 'action -> 'state) -> 'state -> 'state * ('action -> unit)

val useReducerWithMapState : ('state -> 'action -> 'initialState) -> 'initialState -> ('initialState -> 'state) -> 'state * ('action -> unit)

val useEffect : (unit -> (unit -> unit) option) -> unit

val useEffect0 : (unit -> (unit -> unit) option) -> unit

val useEffect1 : (unit -> (unit -> unit) option) -> 'dependency array -> unit

val useEffect2 : (unit -> (unit -> unit) option) -> ('dependency1 * 'dependency2) -> unit

val useEffect3 : (unit -> (unit -> unit) option) -> ('dependency1 * 'dependency2 * 'dependency3) -> unit

val useEffect4 : (unit -> (unit -> unit) option) -> ('dependency1 * 'dependency2 * 'dependency3 * 'dependency4) -> unit

val useEffect5 : (unit -> (unit -> unit) option) -> ('dependency1 * 'dependency2 * 'dependency3 * 'dependency4 * 'dependency5) -> unit

val useEffect6 : (unit -> (unit -> unit) option) -> ('dependency1 * 'dependency2 * 'dependency3 * 'dependency4 * 'dependency5 * 'dependency6) -> unit

val useLayoutEffect0 : (unit -> (unit -> unit) option) -> unit

val useLayoutEffect1 : (unit -> (unit -> unit) option) -> 'dependency array -> unit

val useLayoutEffect2 : (unit -> (unit -> unit) option) -> ('dependency1 * 'dependency2) -> unit

val useLayoutEffect3 : (unit -> (unit -> unit) option) -> ('dependency1 * 'dependency2 * 'dependency3) -> unit

val useLayoutEffect4 : (unit -> (unit -> unit) option) -> ('dependency1 * 'dependency2 * 'dependency3 * 'dependency4) -> unit

val useLayoutEffect5 : (unit -> (unit -> unit) option) -> ('dependency1 * 'dependency2 * 'dependency3 * 'dependency4 * 'dependency5) -> unit

val useLayoutEffect6 : (unit -> (unit -> unit) option) -> ('dependency1 * 'dependency2 * 'dependency3 * 'dependency4 * 'dependency5 * 'dependency6) -> unit

val setDisplayName : 'component -> string -> unit

module Children : sig ... end

val suspend : 'a Lwt.t -> unit

module Experimental : sig ... end

val useTransition : unit -> bool * ((unit -> unit) -> unit)

val useDebugValue : 'value -> ?format:('value -> string) -> unit

val useDeferredValue : 'value -> 'value