Module React.JSX

All of those types are used by the server-reason-react.ppx internally to represent valid React code from the server. It currently different from reason-react-ppx due to a need for knowing the types since ReactDOM needs to render differently depending on the type.

type event =

| Drag of Event.Drag.t -> unit

| Mouse of Event.Mouse.t -> unit

| Selection of Event.Selection.t -> unit

| Touch of Event.Touch.t -> unit

| UI of Event.UI.t -> unit

| Wheel of Event.Wheel.t -> unit

| Clipboard of Event.Clipboard.t -> unit

| Composition of Event.Composition.t -> unit

| Transition of Event.Transition.t -> unit

| Animation of Event.Animation.t -> unit

| Pointer of Event.Pointer.t -> unit

| Keyboard of Event.Keyboard.t -> unit

| Focus of Event.Focus.t -> unit

| Form of Event.Form.t -> unit

| Media of Event.Media.t -> unit

| Inline of string

All event callbacks

type prop =

| Action : (string * string * _ Runtime.server_function) -> prop

| Bool of string * string * bool

| String of string * string * string

| Style of (string * string * string) list

| DangerouslyInnerHtml of string

| Ref of domRef

| Event of string * event

JSX.prop is the representation of HTML/SVG attributes and DOM events

Helpers to create JSX.prop without variants, helpful for function application

val bool : string -> string -> bool -> prop

val string : string -> string -> string -> prop

val style : (string * string * string) list -> prop

val dangerouslyInnerHtml : < __html : string.. > -> prop

val int : string -> string -> int -> prop

val float : string -> string -> float -> prop

val ref : domRef -> prop

val event : string -> event -> prop

module Event : sig ... end